requirements.txt
# Core dependencies
feedparser>=6.0.0
requests>=2.28.2
beautifulsoup4>=4.11.2
python-dateutil>=2.8.2
qdrant-client>=1.4.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
schedule>=1.2.0

# UI dependencies
flask>=2.0.0
flask-socketio>=5.3.0
python-engineio>=4.3.0
python-socketio>=5.7.0
flask-cors>=3.0.10
